refactor(frontend): refactor reading sessions page
