+++ frontend/src/UserPage.tsx
@@ -4,7 +4,7 @@ import UserAvatar from "@/components/UserAvatar"
 import { userApi } from "./api/user"
 import { User, ReadingSession } from "./types/api"
 import { LumiDb } from "./lib/db"
-import { IndividualSessions } from "./ReadingSessionsPage"
+import { IndividualSessions } from "./components/home/readingSessions"
 import { createStore } from "solid-js/store"
 import { readingSessionsApi } from "./api/readingSessions"
 import Spinner from "./components/Spiner"
