+++ frontend/src/pages/home/index.ts
@@ -1,2 +1,3 @@
 export * from "./home.tsx"
 export * from "./library.tsx"
+export * from "./readingSessions.tsx"
